# docs/network-openapi.yaml
openapi: 3.1.0
info:
  title: Smart Glove Network API
  version: 0.1.0
  summary: HTTP and WebSocket wire format used by the current `network` crate.
  description: >
    This spec documents the current payload shape implemented in `src/network/src/`.
    The crate takes concrete endpoint URLs from integration code, so the paths below are
    canonical server-side routes for documentation purposes.

servers:
  - url: https://api.example.com
    description: Canonical HTTPS API origin
  - url: wss://api.example.com
    description: Canonical WebSocket API origin

tags:
  - name: Device Info
  - name: Status Stream
  - name: Machine Control
  - name: Glove Mappings

paths:
  /v1/device-info:
    post:
      tags:
        - Device Info
      summary: Publish device metadata
      description: >
        Called by the device to advertise its `device_id`, `kind`, and the list of event names it
        supports. For `glove` devices, `events` are source events the glove may emit on the status
        stream. For `machine` devices, `events` are control events the machine accepts from the
        server over the machine control stream.
        The current client accepts any `2xx` status and ignores the response body.
      operationId: sendDeviceInfo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceInfoRequest'
            examples:
              default:
                value:
                  device_id: aca704299de8
                  kind: glove
                  events:
                    - gesture.wave
                    - gesture.fist
                    - battery.low
      responses:
        '200':
          description: Device info accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '202':
          description: Device info accepted asynchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '400':
          description: Invalid payload
        '409':
          description: Conflicting device registration state

  /v1/gloves/{glove_device_id}/mappings:
    get:
      tags:
        - Glove Mappings
      summary: List machine mappings for a glove
      description: >
        Returns all remote-control mappings owned by the specified glove. Each mapping links one
        glove-emitted event to one machine control event.
      operationId: listGloveMappings
      parameters:
        - $ref: '#/components/parameters/GloveDeviceIdPath'
      responses:
        '200':
          description: Mappings for the glove
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GloveMappingListResponse'
    post:
      tags:
        - Glove Mappings
      summary: Create a machine mapping for a glove
      description: >
        Registers one mapping from a glove source event to a machine target event. The server
        should validate that the source device exists with `kind=glove`, the target device exists
        with `kind=machine`, and both referenced events were declared in each device's
        `/v1/device-info` payload.
      operationId: createGloveMapping
      parameters:
        - $ref: '#/components/parameters/GloveDeviceIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGloveMappingRequest'
            examples:
              default:
                value:
                  source_event: event.infer.wave
                  target_machine_id: feedfacecafe
                  target_event: machine.led.on
      responses:
        '201':
          description: Mapping created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GloveMapping'
        '400':
          description: Invalid payload
        '404':
          description: Glove or machine not found
        '409':
          description: Duplicate or conflicting mapping

  /v1/gloves/{glove_device_id}/mappings/{mapping_id}:
    delete:
      tags:
        - Glove Mappings
      summary: Delete a machine mapping for a glove
      operationId: deleteGloveMapping
      parameters:
        - $ref: '#/components/parameters/GloveDeviceIdPath'
        - $ref: '#/components/parameters/MappingIdPath'
      responses:
        '204':
          description: Mapping deleted
        '404':
          description: Mapping not found

  /v1/ws:
    get:
      tags:
        - Status Stream
      summary: Upgrade to the device status WebSocket
      description: >
        HTTP upgrade endpoint used by the device for online notifications, discrete events,
        and buffered sensor telemetry. After a successful `101` response, the client sends only
        text frames containing JSON messages matching the schemas documented in
        `x-websocket-messages`.
      operationId: openStatusWebSocket
      responses:
        '101':
          description: Switching Protocols
        '400':
          description: Invalid upgrade request
        '401':
          description: Authentication required
        '403':
          description: Device not allowed to open the stream
      x-websocket-messages:
        direction: client_to_server
        contentType: application/json
        oneOf:
          - $ref: '#/components/schemas/DeviceOnlineFrame'
          - $ref: '#/components/schemas/DeviceEventFrame'
          - $ref: '#/components/schemas/SensorBatchFrame'
        examples:
          online:
            value:
              kind: device.online
              device_id: aca704299de8
          event:
            value:
              kind: device.event
              device_id: aca704299de8
              event: battery.low
              payload:
                percent: 11
          sensorBatch:
            value:
              kind: sensor.batch
              device_id: aca704299de8
              sample_rate_hz: 100
              start_tick_ms: 128340
              samples:
                - imu_acc: [0.11321, 0.42855, 1.00231]
                  vel: [0.01340, 0.20441, -0.03210]
                  flex:
                    thumb: 0.41231
                    index: 0.53311
                    middle: 0.48421
                    ring: 0.39200
                    pinky: 0.31541
                - imu_acc: [0.11648, 0.43002, 1.00188]
                  vel: [0.01512, 0.20395, -0.03181]
                  flex:
                    thumb: 0.41302
                    index: 0.53395
                    middle: 0.48490
                    ring: 0.39266
                    pinky: 0.31602

  /v1/machine-ws:
    get:
      tags:
        - Machine Control
      summary: Upgrade to the machine control WebSocket
      description: >
        HTTP upgrade endpoint used by machine devices for real-time remote control. The machine
        actively connects to this endpoint after provisioning and sending `/v1/device-info`.
        Client-to-server traffic is used for online presence and optional execution results.
        Server-to-client traffic delivers control triggers derived from glove events and glove
        mapping rules.
      operationId: openMachineControlWebSocket
      responses:
        '101':
          description: Switching Protocols
        '400':
          description: Invalid upgrade request
        '401':
          description: Authentication required
        '403':
          description: Device not allowed to open the stream
      x-websocket-messages:
        direction: bidirectional
        contentType: application/json
        client_to_server:
          oneOf:
            - $ref: '#/components/schemas/MachineOnlineFrame'
            - $ref: '#/components/schemas/MachineResultFrame'
          examples:
            online:
              value:
                kind: machine.online
                device_id: feedfacecafe
            result:
              value:
                kind: machine.result
                device_id: feedfacecafe
                request_id: req_123
                status: ok
                payload: null
        server_to_client:
          oneOf:
            - $ref: '#/components/schemas/MachineTriggerFrame'
          examples:
            trigger:
              value:
                kind: machine.trigger
                request_id: req_123
                device_id: feedfacecafe
                event: machine.led.on
                source_device_id: aca704299de8
                source_event: event.infer.wave
                payload: null

components:
  parameters:
    GloveDeviceIdPath:
      name: glove_device_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-f]{12}$'
      description: Device identifier for a registered glove.

    MappingIdPath:
      name: mapping_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
      description: Server-assigned mapping identifier.

  schemas:
    AckResponse:
      type: object
      additionalProperties: true
      properties:
        ok:
          type: boolean
      required:
        - ok

    DeviceInfoRequest:
      type: object
      additionalProperties: false
      properties:
        device_id:
          type: string
          description: Lowercase hexadecimal device identifier derived from the ESP factory MAC.
          pattern: '^[0-9a-f]{12}$'
        kind:
          type: string
          description: Device class.
          enum:
            - glove
            - machine
        events:
          type: array
          description: >
            Event names declared by this device. For `glove`, these are source events the glove
            may emit over `/v1/ws`. For `machine`, these are control events the machine accepts
            over `/v1/machine-ws`.
          items:
            type: string
          minItems: 0
      required:
        - device_id
        - kind
        - events

    DeviceOnlineFrame:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
          const: device.online
        device_id:
          type: string
          pattern: '^[0-9a-f]{12}$'
      required:
        - kind
        - device_id

    MachineOnlineFrame:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
          const: machine.online
        device_id:
          type: string
          pattern: '^[0-9a-f]{12}$'
      required:
        - kind
        - device_id

    MachineResultFrame:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
          const: machine.result
        device_id:
          type: string
          pattern: '^[0-9a-f]{12}$'
        request_id:
          type: string
        status:
          type: string
          enum:
            - ok
            - error
        payload:
          $ref: '#/components/schemas/JsonValue'
      required:
        - kind
        - device_id
        - request_id
        - status
        - payload

    MachineTriggerFrame:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
          const: machine.trigger
        request_id:
          type: string
        device_id:
          type: string
          pattern: '^[0-9a-f]{12}$'
        event:
          type: string
        source_device_id:
          type: string
          pattern: '^[0-9a-f]{12}$'
        source_event:
          type: string
        payload:
          $ref: '#/components/schemas/JsonValue'
      required:
        - kind
        - request_id
        - device_id
        - event
        - source_device_id
        - source_event
        - payload

    CreateGloveMappingRequest:
      type: object
      additionalProperties: false
      properties:
        source_event:
          type: string
        target_machine_id:
          type: string
          pattern: '^[0-9a-f]{12}$'
        target_event:
          type: string
      required:
        - source_event
        - target_machine_id
        - target_event

    GloveMapping:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        source_device_id:
          type: string
          pattern: '^[0-9a-f]{12}$'
        source_event:
          type: string
        target_machine_id:
          type: string
          pattern: '^[0-9a-f]{12}$'
        target_event:
          type: string
      required:
        - id
        - source_device_id
        - source_event
        - target_machine_id
        - target_event

    GloveMappingListResponse:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GloveMapping'
      required:
        - items

    DeviceEventFrame:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
          const: device.event
        device_id:
          type: string
          pattern: '^[0-9a-f]{12}$'
        event:
          type: string
        payload:
          $ref: '#/components/schemas/JsonValue'
      required:
        - kind
        - device_id
        - event
        - payload

    SensorBatchFrame:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
          const: sensor.batch
        device_id:
          type: string
          pattern: '^[0-9a-f]{12}$'
        sample_rate_hz:
          type: integer
          minimum: 1
        start_tick_ms:
          type: integer
          minimum: 0
          description: >
            Device-local monotonic tick in milliseconds when the first sample in this batch was
            captured. Sample time for each item is reconstructed as
            `start_tick_ms + index * (1000 / sample_rate_hz)`.
        samples:
          type: array
          items:
            $ref: '#/components/schemas/BufferedSensorSample'
      required:
        - kind
        - device_id
        - sample_rate_hz
        - start_tick_ms
        - samples

    BufferedSensorSample:
      type: object
      additionalProperties: false
      properties:
        imu_acc:
          $ref: '#/components/schemas/Axis3'
        vel:
          $ref: '#/components/schemas/Axis3'
        flex:
          $ref: '#/components/schemas/FlexReadings'
      required:
        - imu_acc
        - vel
        - flex

    Axis3:
      type: array
      minItems: 3
      maxItems: 3
      items:
        type: number

    FlexReadings:
      type: object
      additionalProperties: false
      properties:
        thumb:
          type: number
        index:
          type: number
        middle:
          type: number
        ring:
          type: number
        pinky:
          type: number
      required:
        - thumb
        - index
        - middle
        - ring
        - pinky

    JsonValue:
      description: Arbitrary JSON value accepted by `StatusReporter::send_event`.
      anyOf:
        - type: object
          additionalProperties:
            $ref: '#/components/schemas/JsonValue'
        - type: array
          items:
            $ref: '#/components/schemas/JsonValue'
        - type: string
        - type: number
        - type: integer
        - type: boolean
        - type: 'null'